+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/MakssatsTilsynBarn.kt
@@ -26,8 +26,13 @@ private val MAX = YearMonth.of(2099, 12)
 val satser: List<MakssatsTilsynBarn> =
     listOf(
         MakssatsTilsynBarn(
+            fom = YearMonth.of(2024, 1),
+            tom = MAX,
+            beløp = mapOf(1 to 4650, 2 to 6066, 3 to 6875),
+        ),
+        MakssatsTilsynBarn(
             fom = YearMonth.of(2023, 7),
-            tom = MAX, // TODO
+            tom = YearMonth.of(2023, 12),
             beløp = mapOf(1 to 4480, 2 to 5844, 3 to 6623),
         ),
         MakssatsTilsynBarn(
@@ -36,3 +41,14 @@ val satser: List<MakssatsTilsynBarn> =
             beløp = mapOf(1 to 4369, 2 to 5700, 3 to 6460),
         ),
     )
+
+fun finnMakssats(måned: YearMonth, antallBarn: Int): Int {
+    val sats = satser.find { måned in it.fom..it.tom }
+        ?: error("Finner ikke satser for $måned")
+    return when {
+        antallBarn == 1 -> sats.beløp[1]
+        antallBarn == 2 -> sats.beløp[2]
+        antallBarn > 2 -> sats.beløp[3]
+        else -> null
+    } ?: error("Kan ikke håndtere satser for antallBarn=$antallBarn")
+}
